requirements.txt
bespokelabs-curator
aryn-sdk